optimizacion de variables

+++ control/include/motor.h
@@ -10,20 +10,20 @@ namespace MotorID{
 }
 class motor_{
 private:
-    int in1;
-    int in2;
+    uint8_t in1;
+    uint8_t in2;
     uint8_t enable;
-    uint32_t speed; 
+    uint16_t speed; 
     int ticsSpeed;
     unsigned long last_time;
 public:
     volatile int tics;
     volatile int deltaTics;
-    void initialize(int,int,int,int);
-    void initEncoder(int);
+    void initialize(uint8_t,uint8_t,uint8_t,uint8_t);
+    void initEncoder(uint8_t);
     motor_();
     void updateTics();
-    void setSpeed(int);
+    void setSpeed(uint16_t);
     void ahead();
     void back();
     void stop();
